+++ CHANGELOG.md
@@ -137,3 +137,4 @@
 * Make `env` adaptable effect and remove `local`
 * Define `Accumulator` effect and it's adaptable `gather` method
 * Define `|->` (Coalgebra) and `<-|` (Algebra) type synonyms
+* Remove `oblige` in favor of `Liftable` instance of `Continuation`